deploy builder
Given an AkkaCluster, fill in defaults as needed and generate ancillary
resources as needed. The results are ordinary kubernetes resource
objects which have both meta and runtime object interfaces.

Currently the use case is to support optional RBAC resources, as well as
various defaults like label selectors and deployment strategies.